feat: rewrites to facilitate removing Bluebird
