+++ .gitignore
@@ -1,3 +1,2 @@
-build/*.js
 coverage/
 node_modules/
