+++ .prettierignore
@@ -1,4 +1,3 @@
-build/
+__tests__/__fixtures__/*
 coverage/
 node_modules/
-tests/fixtures/*
