+++ __tests__/__fixtures__/code/javascript.js
