+++ __tests__/__fixtures__/project/swaggerBase.json
@@ -1,0 +1,6 @@
+{
+  "swagger": "2.0",
+  "host": "petstore.swagger.io",
+  "basePath": "/api",
+  "schemes": ["http"]
+}
