+++ __tests__/__fixtures__/project/swaggerBase.yaml
