+++ __tests__/loader.test.js
@@ -4,7 +4,7 @@ const jsYaml = require('js-yaml');
 const Loader = require('../src/loader');
 
 describe('Loader', () => {
-  describe('.resolvePaths', () => {
+  describe('#resolvePaths', () => {
     it('resolves multiple string paths', () => {
       const inputPaths = [`${process.cwd()}/package.json`, `${process.cwd()}/README.md`];
       return Loader.resolvePaths(inputPaths).then(filepaths => {
@@ -13,7 +13,7 @@ describe('Loader', () => {
     });
 
     it('resolves mixtures of glob and string paths', () => {
-      const inputPaths = [`${process.cwd()}/tests/*.js`, `${process.cwd()}/README.md`];
+      const inputPaths = [`${process.cwd()}/__tests__/*.js`, `${process.cwd()}/README.md`];
       return Loader.resolvePaths(inputPaths).then(filepaths => {
         expect(filepaths.length).toBeGreaterThan(1);
         filepaths.forEach(filepath => {
@@ -32,42 +32,42 @@ describe('Loader', () => {
     });
   });
 
-  describe('.loadBase', () => {
-    it('loads an empty object if no swagger is found', () => {
+  describe('#loadBase', () => {
+    it('loads an empty object if no base definition is found', () => {
       return Loader.loadBase().then(base => {
         expect(base).toStrictEqual({});
       });
     });
 
-    it('loads json file data', () => {
-      const jsonPath = `${__dirname}/fixtures/project/swaggerBase.json`;
+    it('loads JSON file data', () => {
+      const jsonPath = `${__dirname}/__fixtures__/project/swaggerBase.json`;
       return Loader.loadBase(jsonPath).then(base => {
         expect(base.swagger).not.toBeUndefined();
       });
     });
 
-    it('loads yaml file data', () => {
-      const yamlPath = `${__dirname}/fixtures/project/swaggerBase.yaml`;
+    it('loads YAML file data', () => {
+      const yamlPath = `${__dirname}/__fixtures__/project/swaggerBase.yaml`;
       return Loader.loadBase(yamlPath).then(base => {
         expect(base.swagger).not.toBeUndefined();
       });
     });
 
-    it('searches for swagger in the provided directory', () => {
-      const dir = `${__dirname}/fixtures/project/`;
+    it('searches for a base API definition in the provided directory', () => {
+      const dir = `${__dirname}/__fixtures__/project/`;
       return Loader.loadBase(dir).then(base => {
         expect(base.swagger).not.toBeUndefined();
       });
     });
 
-    it('returns an empty object if swagger is not found in a directory', () => {
+    it('returns an empty object if base API is not found in a directory', () => {
       return Loader.loadBase(__dirname).then(base => {
         expect(base).toStrictEqual({});
       });
     });
   });
 
-  describe('.loadFiles', () => {
+  describe('#loadFiles', () => {
     it('loads arrays of files', () => {
       return Loader.loadFiles([`${__dirname}/../package.json`, `${__dirname}/../package.json`]).then(files => {
         expect(files).toHaveLength(2);
@@ -76,14 +76,13 @@ describe('Loader', () => {
     });
 
     it('returns errors for non-existent files', () => {
-      return Loader.loadFiles(['dne.js']).then(files => {
-        expect(typeof files[0]).toBe('object');
-        expect(files[0].message).toContain('no such file or directory');
-      });
+      return expect(() => {
+        return Loader.loadFiles([`${__dirname}/../package.json`, `dne.js`]);
+      }).rejects.toThrow(/no such file or directory/);
     });
   });
 
-  describe('.expandParam', () => {
+  describe('#expandParam', () => {
     it('parses params properly', () => {
       const test = {
         '(query) hi=2* {Integer} This is a description': {
@@ -146,98 +145,98 @@ describe('Loader', () => {
         expect(Loader.expandParam(k)).toStrictEqual(test[k]);
       });
     });
-  });
 
-  describe('.expandParam OAS 3', () => {
-    it('parses params properly', () => {
-      const test = {
-        '(query) hi=2 {Integer} This is a description': {
-          in: 'query',
-          name: 'hi',
-          schema: {
-            type: 'integer',
-            default: 2,
+    describe('OpenAPI 3.0', () => {
+      it('parses params properly', () => {
+        const test = {
+          '(query) hi=2 {Integer} This is a description': {
+            in: 'query',
+            name: 'hi',
+            schema: {
+              type: 'integer',
+              default: 2,
+            },
+            description: 'This is a description',
           },
-          description: 'This is a description',
-        },
-        '(query) hi=2* {Integer} This is a description': {
-          in: 'query',
-          name: 'hi',
-          required: true,
-          schema: {
-            type: 'integer',
-            default: 2,
+          '(query) hi=2* {Integer} This is a description': {
+            in: 'query',
+            name: 'hi',
+            required: true,
+            schema: {
+              type: 'integer',
+              default: 2,
+            },
+            description: 'This is a description',
           },
-          description: 'This is a description',
-        },
-        '(path) hi=2* {Integer} This is a description': {
-          in: 'path',
-          name: 'hi',
-          required: true,
-          schema: {
-            type: 'integer',
-            default: 2,
+          '(path) hi=2* {Integer} This is a description': {
+            in: 'path',
+            name: 'hi',
+            required: true,
+            schema: {
+              type: 'integer',
+              default: 2,
+            },
+            description: 'This is a description',
           },
-          description: 'This is a description',
-        },
-        '(path) hi=2 {Integer} This is a description': {
-          in: 'path',
-          name: 'hi',
-          required: true, // in paths, always require
-          schema: {
-            default: 2,
-            type: 'integer',
+          '(path) hi=2 {Integer} This is a description': {
+            in: 'path',
+            name: 'hi',
+            required: true, // in paths, always require
+            schema: {
+              default: 2,
+              type: 'integer',
+            },
+            description: 'This is a description',
           },
-          description: 'This is a description',
-        },
-        '(body) test {Boolean} this is a description': {
-          in: 'body',
-          name: 'test',
-          schema: {
-            type: 'boolean',
+          '(body) test {Boolean} this is a description': {
+            in: 'body',
+            name: 'test',
+            schema: {
+              type: 'boolean',
+            },
+            description: 'this is a description',
           },
-          description: 'this is a description',
-        },
-        '(body) test {Boolean:hi} this is a description': {
-          in: 'body',
-          name: 'test',
-          schema: {
-            type: 'boolean',
-            format: 'hi',
+          '(body) test {Boolean:hi} this is a description': {
+            in: 'body',
+            name: 'test',
+            schema: {
+              type: 'boolean',
+              format: 'hi',
+            },
+            description: 'this is a description',
           },
-          description: 'this is a description',
-        },
-        '(body) test=true {Boolean} this is a description': {
-          in: 'body',
-          name: 'test',
-          schema: {
-            type: 'boolean',
-            default: true,
+          '(body) test=true {Boolean} this is a description': {
+            in: 'body',
+            name: 'test',
+            schema: {
+              type: 'boolean',
+              default: true,
+            },
+            description: 'this is a description',
           },
-          description: 'this is a description',
-        },
-        '(body) {Boolean:hi} this is a description': {
-          in: 'body',
-          name: '__base__',
-          schema: {
-            type: 'boolean',
-            format: 'hi',
+          '(body) {Boolean:hi} this is a description': {
+            in: 'body',
+            name: '__base__',
+            schema: {
+              type: 'boolean',
+              format: 'hi',
+            },
+            description: 'this is a description',
           },
-          description: 'this is a description',
-        },
-        'test {Boolean:hi} this is a description': false,
-        '(body) test this is a description': false,
-      };
+          'test {Boolean:hi} this is a description': false,
+          '(body) test this is a description': false,
+        };
 
-      Object.keys(test).forEach(k => {
-        expect(Loader.expandParam(k, 3)).toStrictEqual(test[k]);
+        Object.keys(test).forEach(k => {
+          expect(Loader.expandParam(k, 3)).toStrictEqual(test[k]);
+        });
       });
     });
   });
 
-  describe('.loadData', () => {
-    it('loads yaml data', () => {
-      const yamlPath = `${__dirname}/fixtures/project/swaggerBase.yaml`;
+  describe('#loadData', () => {
+    it('loads YAML data', () => {
+      const yamlPath = `${__dirname}/__fixtures__/project/swaggerBase.yaml`;
       const yamlObject = jsYaml.load(fs.readFileSync(yamlPath, 'utf-8'));
       return Loader.loadData(yamlPath).then(yaml => {
         Object.keys(yamlObject).forEach(key => {
@@ -246,8 +245,8 @@ describe('Loader', () => {
       });
     });
 
-    it('loads json data', () => {
-      const jsonPath = `${__dirname}/fixtures/project/swaggerBase.json`;
+    it('loads JSON data', () => {
+      const jsonPath = `${__dirname}/__fixtures__/project/swaggerBase.json`;
       const jsonObject = JSON.parse(fs.readFileSync(jsonPath, 'utf-8'));
       return Loader.loadData(jsonPath).then(json => {
         Object.keys(jsonObject).forEach(key => {
