+++ __tests__/options.test.js
