+++ bin/swagger-inline
@@ -1,0 +1,2 @@
+#! /usr/bin/env node
+require('../src/cli')(process.argv);
