+++ src/options.js
@@ -2,8 +2,7 @@ const path = require('path');
 
 class Options {
   constructor(providedOptions = {}) {
-    // eslint-disable-next-line prefer-object-spread
-    this.options = Object.assign({}, providedOptions);
+    this.options = { ...providedOptions };
 
     Object.keys(Options.DEFAULTS).forEach(option => {
       this.options[option] = this.options[option] || Options.DEFAULTS[option];
