--- tests/fixtures/project/swaggerBase.json
@@ -1,6 +1,0 @@
-{
-    "swagger": "2.0",
-    "host": "petstore.swagger.io",
-    "basePath": "/api",
-    "schemes": ["http"]
-}
